flask/requirements.txt
auto-sklearn
flask
flask-cors
flask-restx
pandas
typing
